refactor: remove unused code and organize things a bit better
+ remove debug logging of restriction registration
